fix(core): keep step phase and voice pitch stable across long runs
- Sequencer::advanceStep receives the full 32-bit uClock counter and
  narrows only after the per-track modulo; truncating to uint8_t first
  aliased every 256 steps onto pattern position 0.
- Voice pitch path reads the injected scale table directly with clamped
  indices and saturated MIDI range; write-only unique-rank caches were
  removed from setScaleTable.
- PitchSnapshot now records the effective scale row so a live
  currentScale change invalidates the static base frequency and
  repitches repeated notes.
- Add unit tests for the wrap phase and voice pitch mapping, and update
  docs/voice.md accordingly.

+++ src/ui/UIEventHandler.cpp
@@ -822,7 +822,7 @@ void clearSequencerStep(Sequencer &sequencer, uint8_t stepIdx)
   }
 }
 
-void advanceSequencerStep(Sequencer &seq, uint8_t current_uclock_step, int mm_distance,
+void advanceSequencerStep(Sequencer &seq, uint32_t current_uclock_step, int mm_distance,
                           const UIState &uiState, VoiceState *voiceState)
 {
   seq.advanceStep(current_uclock_step, mm_distance,
